Move tests to Shared
Move all the code for the tests to a Shared folder for re-use.
