+++ tests/Shared/Constructor.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Martin Costello, 2018. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+namespace MartinCostello.Logging.XUnit;
+
+public enum Constructor
+{
+    ITestOutputHelper,
+
+    IMessageSink,
+}
