+++ tests/Shared/Integration/HttpApplicationTests.cs
